+++ RELEASES.md
@@ -12,6 +12,8 @@ Upcoming
 [Commits](https://github.com/ibissource/iaf/compare/v7.1-B4...HEAD)
 [![Build Status](https://travis-ci.org/ibissource/iaf.png)](https://travis-ci.org/ibissource/iaf)
 
+- Fix NPE in HttpSender when no charset is supplied in multipart response
+
 
 
 7.1-B4